shobarkhamar-complete-backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Async HTTP Client
httpx>=0.27.0

# Image Processing
Pillow>=11.0.0

# Data Processing
numpy>=2.0.0

# Database (PostgreSQL)
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36

# Async SQLAlchemy support
asyncpg>=0.30.0
greenlet>=3.1.0

# Redis and Celery (for async tasks)
redis>=5.2.0
celery>=5.4.0

# Environment Variables
python-dotenv>=1.0.1

# Pydantic Settings
pydantic>=2.10.0
pydantic-settings>=2.7.0

# JWT & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Email validation
email-validator>=2.1.0

# 
gdown
pytest
email-validator
